+++ Sources/KeyboardKit/_Deprecated/LocalizedCalloutActionProvider.swift
@@ -1,0 +1,4 @@
+import Foundation
+
+@available(*, deprecated, message: "This will be removed in KeyboardKit 8.0. Use CalloutActionProvider & LocalizedService instead")
+public typealias LocalizedCalloutActionProvider = CalloutActionProvider & LocalizedService